requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
colorama==0.4.6
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
redis==5.1.0
sqlparse==0.5.1
tzdata==2024.2
